Development/Backend/upload-microservice/requirements.txt
argon2-cffi==25.1.0
argon2-cffi-bindings==21.2.0
blinker==1.9.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
clamd==1.0.2
click==8.2.1
colorama==0.4.6
Flask==3.1.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
minio==7.2.15
packaging==25.0
pycparser==2.22
pycryptodome==3.23.0
PyPDF2==3.0.1
python-dotenv==1.1.1
requests==2.32.4
typing_extensions==4.14.0
urllib3==2.5.0
Werkzeug==3.1.3
